+++ wgsl_to_wgpu/src/bindgroup.rs
@@ -125,6 +125,16 @@ pub fn bind_groups_module(
                         self.set_bind_group(index, bind_group, offsets);
                     }
                 }
+                impl SetBindGroup for wgpu::RenderBundleEncoder<'_> {
+                    fn set_bind_group(
+                        &mut self,
+                        index: u32,
+                        bind_group: &wgpu::BindGroup,
+                        offsets: &[wgpu::DynamicOffset],
+                    ) {
+                        self.set_bind_group(index, bind_group, offsets);
+                    }
+                }
             }
             #set_bind_groups
         }