+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Breaking changes in the generated code will be considered as breaking changes wh
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## unreleased
+### Added
+* Added support for `wgpu::RenderBundleEncoder` for `bind_groups::BindGroups::set` method.
+
 ## 0.13.0 - 2025-01-28
 ### Changed
 * Updated wgpu to latest.